tan 무한대 검출시 int 타입캐스팅 제거.(최초 목적 부동소수점 오차 해결. 90.5도 90으로 취급되는 문제 발생.)

+++ Calculator/src/LogTri/trigonometrical.java
@@ -17,7 +17,7 @@ public class trigonometrical {
 		else if(op.equalsIgnoreCase("tan"))
 
 		{
-			if((int)x % 90 == 0 && (int)x != 0 && (int)x % 180 != 0)
+			if(x % 90 == 0 && x != 0 && x % 180 != 0)
 				result = Double.NaN;
 			else
 			{
